Refactor exception imports and restructure common exceptions
- Updated import paths for exceptions in various modules to use the new common.exceptions module.
- Created a new common.exceptions module to centralize exception handling.
- Added specific exception classes for bot operations and Elaw-related errors.
- Ensured all relevant bots and controllers reference the updated exception classes.
