+++ crawjud/bots/caixa/emissor.py
@@ -13,13 +13,13 @@ from contextlib import suppress
 from time import sleep
 from typing import Self
 
+from common.bot import ClassBot
 from pypdf import PdfReader
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 
 class OtherUtils: ...  # noqa: D101
