+++ crawjud/bots/elaw/andamentos.py
@@ -13,14 +13,14 @@ from contextlib import suppress
 from time import sleep
 from typing import Self
 
+from common.bot import ClassBot
 from selenium.webdriver import Keys
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.wait import WebDriverWait
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 
 class Andamentos(ClassBot):
