+++ crawjud/bots/elaw/complement.py
@@ -19,6 +19,7 @@ from pathlib import Path
 from time import sleep
 from typing import Callable, Self
 
+from common.bot import ClassBot
 from selenium.common.exceptions import NoSuchElementException, TimeoutException
 from selenium.webdriver import Keys  # noqa: F401
 from selenium.webdriver.common.by import By
@@ -26,8 +27,7 @@ from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.wait import WebDriverWait
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 type_doc = {11: "cpf", 14: "cnpj"}
