+++ crawjud/bots/elaw/prazos.py
@@ -15,12 +15,12 @@ import traceback
 from contextlib import suppress
 from typing import Self
 
+from common.bot import ClassBot
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 
 class Prazos(ClassBot):
