+++ crawjud/bots/esaj/busca_pags.py
@@ -10,12 +10,12 @@ from datetime import datetime
 from typing import Self
 
 import pytz
+from common.bot import ClassBot
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 
 class BuscaPags(ClassBot):
