+++ crawjud/bots/esaj/capa.py
@@ -9,12 +9,12 @@ import traceback
 from contextlib import suppress
 from typing import Self
 
+from common.bot import ClassBot
 from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as ec
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 
 class Capa(ClassBot):
