+++ crawjud/bots/esaj/emissao.py
@@ -13,6 +13,7 @@ from time import sleep
 from typing import Self
 
 import requests
+from common.bot import ClassBot
 from pypdf import PdfReader
 from selenium.common.exceptions import TimeoutException
 from selenium.webdriver.common.by import By
@@ -20,8 +21,7 @@ from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.wait import WebDriverWait
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 
 class OtherUtils: ...  # noqa: D101
