+++ crawjud/bots/esaj/movimentacao.py
@@ -11,12 +11,12 @@ from datetime import datetime
 from time import sleep
 from typing import Self
 
+from common.bot import ClassBot
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 
 class Movimentacao(ClassBot):
