+++ crawjud/bots/pje/pauta.py
@@ -19,8 +19,8 @@ from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 
 from crawjud.bots.pje.res.varas_dict import varas as varas_pje
+from crawjud.common.exceptions.bot import ExecutionError
 from crawjud.core import CrawJUD
-from crawjud.exceptions.bot import ExecutionError
 
 
 class Pauta(CrawJUD):
