+++ crawjud/bots/projudi/capa.py
@@ -12,13 +12,13 @@ from datetime import datetime
 from pathlib import Path
 from typing import Self
 
+from common.bot import ClassBot
 from selenium.common.exceptions import StaleElementReferenceException
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 
 class Capa(ClassBot):
