+++ crawjud/bots/projudi/movimentacao.py
@@ -15,6 +15,7 @@ from time import sleep
 from typing import Self
 
 import requests
+from common.bot import ClassBot
 from pypdf import PdfReader
 from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.common.by import By
@@ -22,8 +23,7 @@ from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.ui import Select
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 
 class Movimentacao(ClassBot):
