+++ crawjud/bots/projudi/proc_parte.py
@@ -9,6 +9,7 @@ import traceback
 from contextlib import suppress
 from typing import Self
 
+from common.bot import ClassBot
 from selenium.common.exceptions import (  # noqa: F401
     NoSuchElementException,
     NoSuchWindowException,
@@ -17,8 +18,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from urllib3.exceptions import MaxRetryError  # noqa: F401
 
-from common.bot import ClassBot
-from crawjud.exceptions.bot import ExecutionError
+from crawjud.common.exceptions.bot import ExecutionError
 
 
 class ProcParte(ClassBot):
