+++ crawjud/common/exceptions/bot.py
