+++ crawjud/core/__init__.py
@@ -39,7 +39,7 @@ from addons.logger import dict_config
 from addons.printlogs import PrintMessage
 from crawjud.addons.make_templates import MakeTemplates
 from crawjud.addons.search import SearchController
-from crawjud.exceptions.bot import ExecutionError, StartError
+from crawjud.common.exceptions.bot import ExecutionError, StartError
 from crawjud.types import StrPath, TypeData
 from crawjud.types.elements import type_elements
